requirements.txt
# 기본 Flask 패키지
Flask==3.0.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
SQLAlchemy==2.0.25

# 이미지 처리
Pillow==10.1.0

# 환경 설정
python-dotenv==1.0.0

# 파일 타입 감지
python-magic==0.4.27

# 테스트
pytest==7.4.3
pytest-flask==1.3.0

# 유틸리티
click==8.1.7

# 프로덕션 서버 (선택사항 - 배포 시 설치)
# gunicorn==21.2.0
# eventlet==0.33.3

# 성능 모니터링 (선택사항)
# psutil==5.9.6

# 이미지 최적화 (이미 Pillow로 포함됨)
# 추가 이미지 처리가 필요한 경우:
# opencv-python==4.8.1.78